Make (Abstract)Service extendable
transferred from PR PW#532 by tipsy, branch extendable-spark
