+++ src/test/java/spark/extensions/ExtendedSpark.java
@@ -8,7 +8,7 @@ import spark.Spark;
 public class ExtendedSpark extends Spark {
 
     public static void get(String path, EmptyRoute route) {
-        get(path, route);
+        getInstance().get(path, route);
     }
 
 }
