+++ src/test/java/spark/extensions/ExtensionTest.java
@@ -1,5 +1,7 @@
 package spark.extensions;
 
+import static spark.extensions.ExtendedService.ignite;
+
 import org.junit.Test;
 
 /**
@@ -8,9 +10,12 @@ import org.junit.Test;
 public class ExtensionTest {
 
     @Test
-    public void testExtensions() {
-        ExtendedSpark.get("/", () -> "Hello"); //if it compiles it works ... :)
+    public void testExtensions() { // if it compiles it works ... :)
+        ExtendedSpark.get("/", () -> "Hello");
         ExtendedSpark.stop();
-    }
 
+        ExtendedService extendedHttp = ignite().port(1234).staticFileLocation("/public");
+        extendedHttp.get("/", () -> "Hello");
+        extendedHttp.stop();
+    }
 }
